Cleanup: Moved frequently used function to a utils.py file, and removed unused functions
